refactor: NotificationSystem 분리하기
