+++ src/App.tsx
@@ -1,48 +1,20 @@
 import React, { useState } from "react";
-import { generateItems, renderLog } from "./utils";
-import { useAppContext, AppContext } from "./contexts";
+import { generateItems } from "./utils";
+import { AppContext } from "./contexts";
 import type { AppContextType } from "./contexts";
 import type { User } from "./types";
-import { Header, ItemList, ComplexForm } from "./components";
+import {
+  Header,
+  ItemList,
+  ComplexForm,
+  NotificationSystem,
+} from "./components";
 
 interface Notification {
   id: number;
   message: string;
   type: "info" | "success" | "warning" | "error";
 }
-
-// NotificationSystem 컴포넌트
-export const NotificationSystem: React.FC = () => {
-  renderLog("NotificationSystem rendered");
-  const { notifications, removeNotification } = useAppContext();
-
-  return (
-    <div className="fixed bottom-4 right-4 space-y-2">
-      {notifications.map((notification) => (
-        <div
-          key={notification.id}
-          className={`p-4 rounded shadow-lg ${
-            notification.type === "success"
-              ? "bg-green-500"
-              : notification.type === "error"
-                ? "bg-red-500"
-                : notification.type === "warning"
-                  ? "bg-yellow-500"
-                  : "bg-blue-500"
-          } text-white`}
-        >
-          {notification.message}
-          <button
-            onClick={() => removeNotification(notification.id)}
-            className="ml-4 text-white hover:text-gray-200"
-          >
-            닫기
-          </button>
-        </div>
-      ))}
-    </div>
-  );
-};
 
 // 메인 App 컴포넌트
 const App: React.FC = () => {
