+++ src/components/index.ts
@@ -1,5 +1,6 @@
 import { Header } from "./Header";
 import { ItemList } from "./ItemList";
 import { ComplexForm } from "./ComplexForm";
+import { NotificationSystem } from "./NotificationSystem";
 
-export { Header, ItemList, ComplexForm };
+export { Header, ItemList, ComplexForm, NotificationSystem };
